remove guava from transport and move dependency to new module scalecube utils
