+++ pom.xml
@@ -78,7 +78,8 @@
         <module>benchmark</module>
         <module>examples</module>
         <module>services</module>
-    </modules>
+        <module>Utils</module>
+</modules>
 
     <dependencyManagement>
         <dependencies>
