+++ scalecube-utils/src/main/java/io/scalecube/utils/Preconditions.java
@@ -1,4 +1,4 @@
-package io.scalecube.transport;
+package io.scalecube.utils;
 
 import javax.annotation.Nullable;
 
@@ -22,8 +22,6 @@ public class Preconditions {
    * Ensures the truth of an expression involving one or more parameters to the calling method.
    *
    * @param expression a boolean expression
-   * @param errorMessage the exception message to use if the check fails; will be converted to a string using
-   *        {@link String#valueOf(Object)}
    * @throws IllegalArgumentException if {@code expression} is false
    */
   public static void checkArgument(boolean expression) {
