+++ transport/src/main/java/io/scalecube/transport/Addressing.java
@@ -1,5 +1,7 @@
 package io.scalecube.transport;
 
+import io.scalecube.utils.Strings;
+
 import com.google.common.base.Throwables;
 
 import java.io.IOException;
