+++ transport/src/main/java/io/scalecube/transport/Message.java
@@ -1,6 +1,6 @@
 package io.scalecube.transport;
 
-import static io.scalecube.transport.Preconditions.checkArgument;
+import static io.scalecube.utils.Preconditions.checkArgument;
 
 import java.util.Collections;
 import java.util.HashMap;
