+++ transport/src/main/java/io/scalecube/transport/TransportImpl.java
@@ -1,7 +1,7 @@
 package io.scalecube.transport;
 
-import static io.scalecube.transport.Preconditions.checkArgument;
-import static io.scalecube.transport.Preconditions.checkState;
+import static io.scalecube.utils.Preconditions.checkArgument;
+import static io.scalecube.utils.Preconditions.checkState;
 
 import io.netty.bootstrap.Bootstrap;
 import io.netty.bootstrap.ServerBootstrap;
